+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Changes currently on master, but not under a tag.
 
+### Changes
+- Make BBH/BNS parameter conversion more logical
+- Source frame masses/spins included in posterior
+
 ## [0.3] 2018-01-02
 
 ### Added